Replace \n with newlines in intermediate outputs
Example file:

`commit-linked-issues/SlicerOpenLIFU/#317`:
```
SlicerOpenLIFU	#317	Solution computation failure is silent	When solution computation fails, no error is displayed. The failure is completely silent if one doesn't run the app from a terminal or look in the python console.	bug	arhowe00	CLOSED	2025-05-02T02:12:00Z	2025-05-14T14:55:24Z	@arhowe00
What do you mean by "fails"? If `compute_solution_openlifu` throws an exception?
---
@ebrahimebrahim
> What do you mean by "fails"? If `compute_solution_openlifu` throws an exception?

Yes, for example. Or I guess if it returns a solution that is somehow problematic and causes an exception to be raised while constructing the SlicerOpenLIFUSolution?
---
@arhowe00
If the solution computation fails through a thrown exception, should I forward what the exception says?  It may be the most direct way of conveying information, even though it could be cryptic. The solution computation is pretty much entirely on the python app so it would be difficult for me to parse these reasons out for the benefit of UX
---
@ebrahimebrahim
As just discussed yes forwarding the exception message is a reasonable approach.

This may be a really quick thing also if you add `@display_error` decorator that we sometimes use to an appropriate widget function -- IDK for sure.
---
```